Removed global middleware and made it only apply to a single endpoint
