+++ apps/services/contentful-entry-tagger/src/app/app.module.ts
@@ -1,6 +1,7 @@
 import { LoggingModule } from '@island.is/logging'
-import { Module } from '@nestjs/common'
+import { MiddlewareConsumer, Module, NestModule } from '@nestjs/common'
 import { AppController } from './app.controller'
+import { AppMiddleware } from './app.middleware'
 import { AppRepository } from './app.repository'
 import { AppService } from './app.service'
 
@@ -9,4 +10,8 @@ import { AppService } from './app.service'
   controllers: [AppController],
   providers: [AppService, AppRepository],
 })
-export class AppModule {}
+export class AppModule implements NestModule {
+  configure(consumer: MiddlewareConsumer) {
+    consumer.apply(AppMiddleware).forRoutes('entry-created')
+  }
+}
